+++ behaviortree_future/src/behavior_nodes/mod.rs
@@ -4,5 +4,5 @@ pub use async_action::*;
 mod async_invert;
 pub use async_invert::*;
 
-mod async_sequence;
-pub use async_sequence::*;
+mod async_sequence_or_select;
+pub use async_sequence_or_select::*;
